services/expo-qrcode/requirements.txt
qrcode==7.3.1
Pillow==8.0.1
matplotlib==3.3.2